+++ src/click/termui.py
@@ -119,6 +119,9 @@ def prompt(
                          show_choices is true and text is "Group by" then the
                          prompt will be "Group by (day, week): ".
 
+    .. versionchanged:: 8.3.1
+        A space is no longer appended to the prompt.
+
     .. versionadded:: 8.0
         ``confirmation_prompt`` can be a custom string.
 
@@ -138,10 +141,10 @@ def prompt(
         try:
             # Write the prompt separately so that we get nice
             # coloring through colorama on Windows
-            echo(text.rstrip(" "), nl=False, err=err)
-            # Echo a space to stdout to work around an issue where
+            echo(text[:-1], nl=False, err=err)
+            # Echo the last character to stdout to work around an issue where
             # readline causes backspace to clear the whole line.
-            return f(" ")
+            return f(text[-1:])
         except (KeyboardInterrupt, EOFError):
             # getpass doesn't print a newline if the user aborts input with ^C.
             # Allegedly this behavior is inherited from getpass(3).
@@ -214,6 +217,9 @@ def confirm(
     :param err: if set to true the file defaults to ``stderr`` instead of
                 ``stdout``, the same as with echo.
 
+    .. versionchanged:: 8.3.1
+        A space is no longer appended to the prompt.
+
     .. versionchanged:: 8.0
         Repeat until input is given if ``default`` is ``None``.
 
@@ -231,10 +237,10 @@ def confirm(
         try:
             # Write the prompt separately so that we get nice
             # coloring through colorama on Windows
-            echo(prompt.rstrip(" "), nl=False, err=err)
-            # Echo a space to stdout to work around an issue where
+            echo(prompt[:-1], nl=False, err=err)
+            # Echo the last character to stdout to work around an issue where
             # readline causes backspace to clear the whole line.
-            value = visible_prompt_func(" ").lower().strip()
+            value = visible_prompt_func(prompt[-1:]).lower().strip()
         except (KeyboardInterrupt, EOFError):
             raise Abort() from None
         if value in ("y", "yes"):