+++ CHANGES.rst
@@ -11,6 +11,8 @@ Unreleased
     the ``Context.invoke()`` method. :issue:`3066` :issue:`3065` :pr:`3068`
 -   Fix conversion of ``Sentinel.UNSET`` happening too early, which caused incorrect
     behavior for multiple parameters using the same name. :issue:`3071` :pr:`3079`
+-   Fix rendering when ``prompt`` and ``confirm`` parameter ``prompt_suffix`` is
+    empty. :issue:`3019` :pr:`3021`
 
 Version 8.3.0
 --------------